Move legacy tests
